component voor aanmeldennieuwsbrief gemaakt, deze werkt wel, email logt nu naar local storadge
